fix(core): 修正BlockPos类型从vec3到ivec3

+++ src/Minecraft/game/minecraft/world/level/block/BlockPos.hpp
@@ -1,3 +1,3 @@
 #pragma once
 #include "glm/glm.hpp"
-using BlockPos = glm::vec3;
+using BlockPos = glm::ivec3;
